requirements.txt
# File ini digunakan pip untuk instalasi package python yang diperlukan
# Penggunaan: pip install -r requirements.

#Packages
flask
flask_mysqldb
python-dotenv #.env file

